requirements.txt
python-dotenv
discord.py==2.3.2
aiohttp
